update all double as decimal and add a rounding on the gains in euros report field

+++ RevoProfit.Core/Crypto/Models/CryptoAsset.cs
@@ -3,9 +3,9 @@
 public class CryptoAsset
 {
     public string Jeton { get; init; }
-    public double MontantEnEuros { get; set; }
-    public double Montant { get; set; }
-    public double Frais { get; set; }
+    public decimal MontantEnEuros { get; set; }
+    public decimal Montant { get; set; }
+    public decimal Frais { get; set; }
 
     public override string ToString()
     {
